pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webpost"
version = "0.1.0"
description = "Scenario-based web post runner"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "requests==2.32.3",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
  "redis==5.1.1",
  "rq==1.16.2",
  "python-dotenv==1.0.1",
  "loguru==0.7.2",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["application*", "domain*", "infrastructure*"]
